pyproject.toml
[tool.poetry]
name = "maths_mark"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
vaderSentiment = "^3.3.2"
numpy = "^1.22.3"
keras = "2.8.0"
cv = "^1.0.0"
imutils = "^0.5.4"
pandas = "^1.4.2"
opencv-python = "^4.5.5"
tensorflow = "^2.8.0"
python-multipart = "^0.0.5"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
